Fix windows installation
Run pip to properly install the package with all the metadata.

The extension is still built with cmake as before, but a special
PrebuiltExtension is used to copy it inside the Python package.

Inspired by https://github.com/tim-mitchell/prebuilt_binaries/blob/main/prebuilt_binaries.py

This is a workaround. Will try to do a proper fix in pytango for 9.4.0.

Fix #6